requirements.txt
environs==11.0.0
loguru==0.7.2
python-telegram-bot==13.15
requests==2.31.0
